<Refactor> : image 저장 함수 수정 ---------------------------------------- 1. html에서 img 소스 변경하도록 수정 2. 컨텐츠 수정에도 사용가능하도록 image 저장 함수 수정
----------------------------------------

+++ app/core/image.py
@@ -3,6 +3,7 @@ import json
 from datetime import datetime
 import shutil
 from dotenv import load_dotenv
+from bs4 import BeautifulSoup
 
 load_dotenv(verbose=True)
 class Image():
@@ -88,38 +89,56 @@ class Image():
             return None
         return target_path
 
-    def save_image(self,temp_path,save_path,data,id=None):
+    def save_update_image(self,temp_path:str,save_path:str,data:str,id:int=None,mode:str="su"):
         """
-        data에서 실제 저장되는 사진들만 temp_path에서 save_path로 이동하고 나머지 temp_path는 삭제
+        data에서 실제 저장되는 사진들만 temp_path에서 save_path로 이동하고 나머지 temp_path는 삭제하거나
+        data가 수정되어 save_path의 사진들중 필요없는 사진을 삭제
         
         - temp_path : 사진 임시 저장소
         - save_path : 사진 실제 저장소
-        - data: textedit의 json형식
+        - data: 프론트엔드의 textEditor에서 작성된 html 문자열
         - id : DB에 저장된 data의 id값
+        - mode : 's' - 새로운 이미지를 추가, 'u' - 기존의 이미지들중 필요없는 이미지를 삭제
         """
-        #저장된 data에서 쓰인 사진만 추출 및 텍스트 에디터의 사진 경로를 실제 사진 경로로 수정
-        jsonObject = json.loads(data)
-        imagelist=[]
-        for entity in jsonObject.get("entityMap").values():
-            image_params=entity.get("data").get("src").split('/')
-            imagename=image_params[-2]
-            imagelist.append(imagename)
-            entity["data"]["src"]="/".join(image_params[:-1])
-            if id:
-                entity["data"]["src"]+="?id="+str(id)
-        new_data=json.dumps(jsonObject)
 
-        if not os.path.exists(save_path):
-            os.makedirs(save_path)
-                
-        #temp폴더에서 실제로 저장되지 않은 사진 삭제 및 실제로 쓰인 사진를 문제 id에 맞는 경로로 이동
-        tempimagelist=os.listdir(temp_path)
-        for i in tempimagelist:
-            if i.split(".")[-1]!="keep" and not i in imagelist:
-                os.remove(os.path.join(temp_path,i))
-            else:
-                shutil.move(os.path.join(temp_path,i),os.path.join(save_path,i))
-        os.rmdir(temp_path)
+        #mode값이 틀리면 에러발생
+        for i in mode:
+            if i!="s" and i!="u":
+                raise Exception("mode값은 's','u'값이어야합니다.")
+            
+        imagelist=[]
+        soup = BeautifulSoup(data, 'html.parser')
+        img_tags=soup.find_all('img')
+        # 저장된 data에서 쓰인 사진만 추출 및 텍스트 에디터의 사진 경로를 실제 사진 경로로 수정
+        for i in range(len(img_tags)):
+            image_url=img_tags[i]["src"].split("/temp")[0]
+            img_tags[i]["src"]=image_url
+            imagelist.append(image_url.split("/")[-1])
+            if id:
+                img_tags[i]["src"]+="?id="+str(id)
+        new_data=str(soup)
+
+        if "s" in mode:
+            if not os.path.exists(save_path):
+                os.makedirs(save_path)
+                    
+            #temp폴더에서 실제로 저장되지 않은 사진 삭제 및 실제로 쓰인 사진를 문제 id에 맞는 경로로 이동
+            if os.path.exists(temp_path):
+                temp_imagelist=os.listdir(temp_path)
+                for i in temp_imagelist:
+                    if i.split(".")[-1]!="keep" and not i in imagelist:
+                        os.remove(os.path.join(temp_path,i))
+                    else:
+                        shutil.move(os.path.join(temp_path,i),os.path.join(save_path,i))
+                os.rmdir(temp_path)
+        
+        if "u" in mode:
+            #save폴더에서 data에 더이상 쓰이지 않는 이미지 삭제
+            if os.path.exists(save_path):
+                save_imagelist=os.listdir(save_path)
+                for i in save_imagelist:
+                    if not i in imagelist:
+                        os.remove(os.path.join(save_path,i))
 
         return new_data
